Orgnanize experimental package API code
Move experimental package API views to their own files, as opposed to
having a single large file
